Add blank type literals for GraphQL types (#647)
We have to manually write out GraphQL type literals for "blank" values a lot, e.g. for blank forms and tests and such, which gets annoying fast.  We can actually auto-generate these from the GraphQL schema, though, so we'll do that now.

This uses the new feature to auto-generate a blank value for `AllSessionInfo`, which we can use in tests, rather than having to manually specify it in our test suite and then update it whenever we change the session type.