Auto-generate GraphQL for mutations with session subsets (#650)
This adds support for a `sessionFields` option to mutations in the autogen config, which allows us to specify only specific fields of the `SessionInfo` type to include in mutation responses, to help reduce payload size (which we've been doing when manually writing GraphQL queries).

We do this by adding another configuration option to the `types` section of the autogen config called `includeOnlyFields`, which allows us to provide a list of fields to exclusively include in a specific type.  Then, whenever we encounter a mutation with `sessionFields` set, we create a new context that overrides the `SessionInfo` type configuration to include only the fields specified by `sessionFields`.