Deploy to Heroku on test success (#730)
Currently, I'm manually deploying from my own system using the deploy.py script in the root of the repository. It would be nice if this was done automatically, which this attempts to do.

Right now it's set up to deploy to our dev instance at https://tenants2-dev.herokuapp.com/ whenever changes are made to master, but we can eventually figure out a way to push to prod via this mechanism too.

I'm a bit bummed that Docker Layer Caching (DLC) is a premium feature on CircleCI, which basically means that deploys are going to take a lot longer than they ought to right now. Ah well.