Add a /help page. (#787)
This adds a help page at /help which invites the user to activate compatibility mode. If the user is already in compatibility mode, it suggests the user email our support address.

I'm adding this because there currently isn't a way to activate compatibility mode in a completely manual way. At the very least, if we're helping out a user one-on-one and think that compatibility mode might help things, we should be able to do so. But we should also eventually expose this URL as a menu option so that our users (the tenacious ones, at least) can discover and use it themselves.