Use AWS SQS broker and Django DB result backend with Celery (#821)
In #814 I added optional Celery integration with Redis as the broker, but this proved to be unreliable due to celery/kombu#1019. Then I attempted to switch to RabbitMQ in #819, but this proved prohibitively expensive, so now I'm trying AWS SQS. And because SQS can't be used as a result backend, I'm using django_celery_results, which uses Django's ORM (and in our case, Postgres) for the result backend.