Attempt to use AMQP again, but with CloudAMQP's settings. (#823)
I originally tried using AMQP in #819 but this used an exorbitant number of messages. However, I just found a CloudAMQP post about running Celery which includes some settings that reduce the number of messages processed by RabbitMQ. Given the fact that our current broker, AWS SQS, is labeled as "experimental" in some places and might not be very reliable, we're switching back to AMQP, but with vastly lower message use.