Render React views via an HTTP server (#839)
This makes our server-side rendering work more like an actual lambda/FaaS service by running the lambda process as a long-lived server that communicates with the Django app over HTTP (rather than over stdio as in #409).  This will ensure that requests are served more efficiently, since a brand-new node process won't need to be spawned for each request.

It is currently disabled by default, but can be enabled via the `USE_LAMBDA_HTTP_SERVER` setting. Once we're confident that it works well, we can switch it over by defaulting the setting to `True` (or by simply removing the alternative `LambdaPool` interface).