Add a NotifyServerOfNewBuildPlugin webpack plugin. (#913)
One annoying thing about developing the front-end right now is that the Django server doesn't restart the lambda process until the user reloads the page (at which point the Django server notices the lambda process has changed and restarts it).  This is annoying, though, because the lambda server takes a noticeable time to start up.

I tried adding an auto-reloader thread to the Django server that constantly polled `lambda.js` for changes and reloaded it when it noticed them, but this frequently caused errors because the file is so big that it would sometimes try to restart the process while the file was still being written. This would necessitate having to write something akin to chokidar's `awaitWriteFinish` functionality, which would be both error-prone and not very responsive.

So instead, I added a new Webpack plugin that just pings the Django server whenever a new lambda server build is emitted.  This obviates the need for any kind of polling and is fairly straightforward--the Webpack plugin just needs to know what the URL of the Django server is, which we provide through an environment variable in `docker-compose.yml`.  And if the environment variable isn't defined, no biggie, we just don't ping anything.