Never mark 'urgent and dangerous' in HP forms. (#926)
In practice, the city always wants the "urgent and dangerous" field in HP Action forms to be left unchecked, so that HPD inspections occur (and so the HPD inspection form is part of the HP Action packet). Here we disregard what the user has answered and always set the value to false.