Don't prevent default if opening outbound link in a new tab. (#930)
This (hopefully) fixes #929 by not preventing default link-click behavior for outbound links that will open in a new tab/window. It also sends the GA link click event even if a modifier key is pressed (so we're basically assuming that if the user clicks a link with a modifier key, it's for the purpose of opening it in a new tab, or viewing it in some other way).