Track tracking number for letter of complaint (#925)
This stores the tracking number of the letter of complaint in the `LetterRequest` model and exposes it in the Django admin.  It also adds a "letter sent at" field that stores the date/time the letter was mailed. Finally, we also provide a "rejection reason" field in case the user wants us to mail the letter, but we don't want to do it for some reason.

When a letter is mailed via Lob, the tracking number and sent-at fields are automatically populated.  However, we also allow the fields to be manually populated via the admin UI, if for some reason the user needs to mail a letter manually (i.e. by printing out the letter and walking to the post office).

When the tracking number field is populated, on model save, an SMS will be sent to the user and they will be added to a LOC follow-up campaign, if one has been configured.  This will ultimately replace a Rube-Goldbergian Airtable-Zapier-TextIt machination that currently exists outside this codebase.

Finally, the "Letter mailing needed?" column of the user list admin view now takes the tracking number into account.  Until now, this column was just a reflection of whether the user wanted us to mail their letter for them; however, now it also includes consideration of whether the tracking number for the letter is blank, and if there is no rejection reason: if it is, then we need to mail it (so the column will have a green checkmark in it), otherwise we don't (so the column will have a red X in it).