Store rental history requests in the database (#937)
We're not currently keeping track of who is asking for rent history in any way, and it'd be nice to do this for the user dashboard (e.g. telling them they've already done this), outcome-tracking, and analytics purposes (even if we eventually delete the PII after some period).

This adds a database model to track the information. It also adds an admin view and allows Outreach Coordinators to access it.