Upgrade to Django 2.2.8 (#941)
This upgrades us to the latest security patched version of Django 2.2.

As explained in the 2.2.8 release notes, the Django admin privilege escalation vulnerability was fixed with a breaking change. I noticed that this caused Outreach Coordinators to no longer have access to view/change a user's issues and custom issues, so I fixed this by giving them explicit permission to do so (this is a good thing, I was previously confused by the fact that they could do so without explicit permission).