Resync airtable on LOC mailing & HP Action editing. (#948)
Currently when we send LOCs via the admin interface, the user's changes don't instantly show up in airtable, which is annoying because our admins currently use airtable to manage things a lot. This fixes that.

I also noticed that we don't sync w/ Airtable when HP Actions are edited, so I fixed that too.

Note that I did consider adding a Django signal handler on model save that resynced w/ airtable but decided against this because I didn't know how often models might be saved outside of admin (it might spam airtable) and also because it wouldn't necessarily always fire when we wanted it to (for example, the LOC mailer saves the letter request, not the user object).