Show single row per custom issue. (#952)
This adds a `singleRow` boolean prop to our `<FormsetItem>` component that shows the row's field(s) with the "Delete" checkbox on a single row.  It also adds a `fieldProps` field to textual form fields that allow for arbitrary HTML attributes on the `<div class="field">` that wraps textual form fields.

Using these two new features we can limit the width of the custom issue fields to constrain them to the maximum number of characters, and we can also show them more compactly on a single row per item.