Set HTML5 max length for custom issues. (#955)
This enforces a max length for custom issues using HTML5 form validation's `maxlength` attribute.

My main concern here is that because many existing users have custom issues with up to 250 characters of text (possibly more, if they are really old users), we don't want browsers to truncate the user's data without their consent, thereby losing data.  But after testing this out on IE11, Chrome, and Firefox, all browsers appear to allow the field's initial value to be greater than the max length--they just only allow content to be _removed_ from the field until it is below the max length.  So we should be good.