Remember latest DDO address, prepopulate onboarding/RH w/ it (#960)
This pre-populates onboarding and rental history's "address" fields with the latest address used in DDO, to save users some typing when they follow calls-to-action that involve entering one's address.

This is done via a new `BrowserStorage` class that uses `window.sessionStorage` under the hood.  We're doing everything browser-side because that's actually the most privacy-preserving: if we store the information in the Django session, it will show up if the user closes the browser tab and opens a new one, while storing it in `sessionStorage` means it will go away when the user closes the tab (which they will likely do if they're using a public computer).

Because pre-filling form fields with recently-entered values is non-essential to the functioning of the site (it's just a convenience), we can treat it as a progressive enhancement: it won't be present if the user is in compatibility mode, or if they have JS disabled in their browser.