restructured generic page style to have footer peek just above the fold (#1008)
This styling update adds a new Sass variable called `$footer-offset` that allows us to define how much of the page footer we want to appear above the fold on all pages. A new `min-height` property on all #main divs takes this variable into account, along with the default height for Bulma navbars, to set a minimum height for all pages' main content.