Add a conversations dashboard (#1010)
This adds a conversations dashboard at `/admin/conversations/`, which works much like Google Voice's messaging UI, or the tablet version of an SMS app: contacts appear on the left sidebar, with the conversation for the currently selected contact appearing on the right panel.

The dashboard is powered by a new `texting_history` Django app, which has a single model called `Message`.  Instances of this model are populated via a new `update_texting_history` management command, which talks directly to Twilio to retrieve SMS history.  A new admin-only GraphQL mutation has also been added to do the same thing, which allows the dashboard's front-end to show new updates in almost-real-time, as it just executes this mutation at regular intervals (currently every 3 seconds).

Because the messaging history is pulled directly from Twilio, and because our JustFix phone number is currently shared between this codebase and TextIt, not all contacts will exist in our user database.  Ones that don't will only show up as phone numbers, rather than the contact's full name, and the top of the conversation panel will say "This phone number does not seem to have an account with us."

The dashboard also allows searching for users via their phone number or name.  Messages which couldn't be delivered are shown with a ❌ on the left-hand sidebar and red text on the right side.

This also marks the first addition of admin code to the React front-end.  Admin UX and code has different assumptions than that for other users, and I've documented them at `frontend/lib/admin/README.md`.