Store our own phone number in texting history. (#1018)
We just discovered that we (apparently completely erroneously) initiated some TextIt flows from a completely different phone number for a period of time.  To accommodate for this, we're adding an `our_phone_number` field to the `texting_history.Message` model.  We should be able to store the other number's messages by temporarily changing the value of `TWILIO_PHONE_NUMBER` and running `update_texting_history`.  Because our SMS dashboard just retrieves all text messages and sorts them by the date sent/ordering, the two numbers' messages should be seamlessly "woven" into a single conversation thread with the end-user (even though they didn't really experience it that way, this is the easiest thing for us to manage).

We don't technically _need_ to store the phone number we sent the message from in our database, but I wanted to, in order to preserve data provenance.