Add a "Lint and typecheck Python" GitHub action. (#1020)
Frequently our CircleCI errors fail due to linting/typechecking-related errors, but we're often notified of them minutes after we've pushed a change because of how long it takes for CircleCI to initialize everything, and where the linting/typechecking currently occurs in the test runs.  Also, when CircleCI fails, we don't know what caused the failure, and we have to sift through all the output to see what's wrong; rather than one big mega-check that tests everything, it might be useful to migrate to having a variety of checks that each test specific parts of the codebase, so that we have a better idea of where a problem is when something fails (it also potentially means shorter CI time because we can run these checks concurrently).

This uses GitHub Actions to run Python's flake8 and mypy as quickly as possible--in under 40 seconds, though it's usually more like 30 or even less--so that we're alerted to such common problems sooner rather than later.  The CircleCI task still also runs these tools, because we really don't want to accidentally push anything to production that violates their checks, so right now this new check is a redundant "early warning" more than anything else.

Note that I chose to use GitHub Actions here instead of CircleCI because as far as I can tell, CircleCI's free plan gives us a very low number of concurrent jobs (it seems like they only allow for one at a time), while GitHub Actions allows for a maximum of 20 for its free plan.