Disable analytics for staff accounts. Fixes #1030. (#1031)
This disables analytics (FullStory, Facebook Pixel, GTM, GA, etc) for all users that are staff. This is done largely as a security precaution to fix #1030, so that potentially sensitive administrative information isn't accidentally leaked to analytics services. However, it also lays the foundation for allowing any user to opt-out of analytics if they want.