Don't raise GraphQL errors on auth failure (#1034)
This fixes #1032 by not propagating auth failures as GraphQL errors.

In short, going forward, we only want to raise GraphQL errors that are the equivalent of 500 internal server errors. Anything that is the equivalent of a HTTP 4xx should be returned as an error response instead, which seems to be an accepted convention as far as I can tell (and it's what we're doing with form mutations too).