Upgrade to Apollo 2.25. (#1040)
This upgrades us to the latest version of Apollo, which appears to speed up the generation of TypeScript files.

It looks like the only difference in the TS output is that Apollo is adding a `// @generated` comment at the top of each generated file.  For reference, I was able to discern this by temporarily commenting out the lines in `.gitignore` pointing to anything in `frontend/lib/queries`, staging that directory, then upgrading Apollo and re-running `node querybuilder.js` to regenerate the files in it with the new version of Apollo.