Add basic DocuSign integration. (#1043)
This adds basic DocuSign integration.  We're using their JSON Web Token (JWT) method of authentication because the sending of "envelopes" (what DocuSign calls documents to be signed + associated metadata) on behalf of a DocuSign user will be fully automated, which requires us to impersonate them from our end.

DocuSign authentication is really complicated; the short story is that three configuration parameters are required from the environment, in addition to a private key stored in the database, as well as a consent code that is obtained via an OAuth flow and also stored in the database.  Running `manage.py testdocusign` will guide administrators through the process via hopefully-helpful `ImproperlyConfigured` error messages.

I've also added a *temporary* proof-of-concept signing flow for the HP Action documents at `/hp/docusign/`, which allows a user who has created an HP Action packet _and_ has an email address configured to go through a basic signing process.  None of the view code is particularly well-tested because it will be replaced by a GraphQL mutation pretty soon.