Add Eviction Moratorium banner to the Tenant Platform (#1047)
This PR adds a banner announcing JustFix's status during the COVID-19 pandemic and links to RTC's FAQ page on the Eviction Moratorium.

I essentially created a front-end component that gets added to the AppWithoutRouter component— essentially rendering the banner below the nav bar on any page. I also included a "close" button that removes the banner, and it seems that the state of the banner stays consistent throughout at user session, so users only need to close the banner once to keep it hidden for their whole session on the Tenant Platform.