Add email to onboarding and require password (#1048)
This modifies our onboarding flow:

* It adds a required "email address" field to step 4.
* It makes the password required, rather than optional.
* It sends a verification email once the user has registered.
* It adds a fifth screen to the flow, which thanks users for signing up and tells them about the verification email.

The back-end changes are implemented via new GraphQL mutations, to ensure that any users still on the old onboarding flow (and old front-end code) won't be disrupted.  We should eventually remove the deprecated mutations once we're confident that everyone is on the new flow, though.