Add scaffolding for emergency HP action flow (#1055)
This creates an alternative emergency HP action flow under `/ehp` (as opposed to `/hp`, which houses the regular HP action flow).  It's only enabled if DocuSign integration (#1043) is also enabled.

This involved a bunch of refactoring of the regular HP action flow, to ensure that we could easily reuse common functionality.

Finally, I'm globally applying the React `act()` workaround introduced in #760, because it gets triggered whenever we use React Hooks, which at this point will be _all the time_.  I don't even know of any situations where we've gotten a _non_-spurious `act()` error, so I think this should be fine, at least until we upgrade to React 16.9 and should be able to remove this workaround (see #762).

The one flaw with this scaffolding right now is that after completing onboarding, users will be taken to the regular HP action flow because I haven't added an emergency HP action intent that's separate from the regular one.  Will figure that out later.