Implement emergency HP issue checklist (#1058)
This implements the emergency HP issue checklist!  It works by creating a `HPActionDetails` model configured to sue for repairs and not harassment (both must have boolean values for the LHI endpoint to generate a form), as well as syncing a smaller set of issues (no heat, no hot water) to the user's regular issue checklist.  Right now this means that if the user had 100 issues from LOC, they will still generate an emergency HP form that has tons of issues in it, but we can deal with that situation later.