Add DocuSign integration to EHP flow. (#1059)
This hooks up our existing DocuSign plumbing, introduced in #1043, to the Emergency HP Action flow, allowing users to go through the whole process from onboarding to e-signing.

We're only currently supporting the "happy path" where users are assumed to have an email address, and we're not even checking that it's validated yet, to make it easier to test for now.  We also don't have any tests for the new functionality yet either; that will come soon.

This also removes the temporary views introduced in #1043 to demo the HP signing, since they're redundant given the new functionality.