Add PyPDF2, remove instructions from emergency HPA packet (#1066)
This adds PyPDF2 and uses it to optionally remove the first two pages of instructions in LHI's generated HP Action PDF.  Right now we're calling this the "emergency HP Action PDF" because we might end up doing more than just removing the instructions (for instance, we might move the blurbs added in #1060 to this step, rather than having them done in DocuSign).

Normally a user's HPA can be viewed at `/hp/latest.pdf`.  If `?em=on` is added to the end of that, their emergency HPA will show up instead.  This is a quick and easy way of testing things out in development.

Note that this doesn't add tests for some of the edge cases; will have to handle that in a future PR due to time constraints (see #1068).