Add eslint and enable its react hooks plugin. (#1077)
This addresses #495 by adding eslint and enabling *only* the official React eslint hooks plugin.

It seems we're already using it in some ways that eslint disapproves of, but because some of them touch code that seems to have been working fine for months, I've disabled it on several lines.  What's most important right now is that we use this extremely valuable plugin going forward.