Remove pauseForModalFocus() test utility. (#1083)
The really weird thing about this is that I thought I'd have to fix a bug in react-aria-modal's focus trap to get rid of this function, but I didn't. It's possible that upgrading to the newest version of jest (which updated JSDOM too) might have obviated the need to change it.