Add a useBrowserStorage() hook and use it. (#1084)
This resolves part of #1073 by adding a useBrowserStorage() React Hook which simplifies the use of browser storage in a progressively-enhanced context. In short, it works similarly to useState(), only it pulls from/updates browser storage, and the initial value pre-mount is always a default value to ensure that the initial render on server and client are always identical.