Include email address in email verification signing payload (#1097)
We're going to be providing a UI to allow users to change their email after we've already sent a verification email to them, in case they realize they made a typo or somesuch. However, this opens up a security hole in our current workflow (which was taken from django-registration): because the payload we include in the verification link we send them includes only their username, it's entirely possible for them to enter an email address they own, then change their email address to one that they don't own, and then click the old verification link. Not good!

This changes things around so that we encode and sign both their username and their current email address in the verification link, which means that if they try to pull the aforementioned shenanigan, it won't work: upon clicking the old link, the back-end will realize that the user's current email no longer matches the email embedded in the verification link, and refuse to verify the user.

The one downside of this approach is that it means longer verification URLs, but they're not horrible (the code embedded in them will probably average around 90-100 characters). And anyways, once we start sending HTML emails, very few people will actually see the full link.