Verify user's email in EHP flow (#1102)
This requires the user to verify their email address before signing an Emergency HP Action. If the user doesn't have an email address yet, we ask them for it first (we also allow the user to change their email if e.g. they mis-typed it during onboarding).

For expediency of implementation, the force-verification is in a somewhat non-ideal place right now: right before the landlord info screen, because that's the latest screen at which we have a "next" button that simply moves to the next step of the flow, rather than doing something big like generating a PDF or starting a signing process.

Fixes #1085. Fixes #1086.