Factor out MailingAddress base model, add state choices (#1111)
This factors out all the mailing address-related fields and functionality from our `AddressDetails` model in loc (added in #561) into a Django abstract model so it can easily be reused in other models, which will likely be needed for #1092.

It also adds a list of U.S. states and territories to validate the "state" field.  This list is in `common-data` so it can be reused by the front-end for #1092 in the near future.

The included migration is trivial, as it just updates the help text and choices for the `state` field.