Add granular address info to NychaOffice model. (#1112)
One of the blockers to #959 and most likely #1113 is that the way we're storing NYCHA office address information is "blobby", but the original CSVs we're pulling from actually contain more granular information.  This copies over the granular information so once we move to a granular model, we can use the more granular address data.

Actually populating the granular fields will require re-running the `loadnycha` management command.