Use granular address info for landlords. (#1117)
This changes our underlying data models to store more granular details about landlord mailing addresses when possible (e.g., city/state/zip as separate fields, rather than as a single blob of text).  The old "single blob of text" address will now be used as a legacy fallback.

It does _not_ modify the front-end to actually ask users for the granular information yet.  As such it tees us up for #959 and #1092 but doesn't actually fix them.

In fact the only tangible benefit this code has right now is to alleviate the need for the `AddressDetails` model introduced in #561 for cases where the landlord information was pulled directly from open data: because we'll be delivering the more granular information directly to Lob, this should reduce the chance for Lob to mis-parse the address and regard it as undeliverable.