Allow email to be verified while at start of verification. (#1129)
In EHPA, if the user hasn't yet clicked on their verification email sent during onboarding, they are eventually shown a page that allows them to re-send their verification email.  However, if they manage to find their email and click on it, the page would stay the same, requiring them to re-send another email to themselves in order to move forward.

This adds a session poller which automatically moves them to the success page if they clicked the verification link while on the start page.

This also adds the "is verified email" checkbox to the Django admin ,to make all this easier to manually test (and obviously to make life easier for actual admins too).