Add DocusignEnvelope model to track signing status (#1132)
This adds a `DocusignEnvelope` model to help us persistently track the signing status of the user's Emergency HP Action.  It also ensures that only one DocuSign Envelope is created per set of HP Action documents generated, which reduces DocuSign costs.  We also use the status to determine whether to send the user to the EHP confirmation page, fixing #1115.