Don't follow-up a EHP with normal HP text messages. (#1135)
Currently, generating (but not signing) emergency HP forms will result in sending the user the same text message as a normal HP form, and initiating them on the normal HP action follow up flow. Both of these are bad because they make mention of filing papers in-person in court!

This ensures that doesn't happen.