Poll for email verification every 2.5s, don't interrupt user. (#1136)
This improves the email verification flow by polling the server to see if the email has been verified every 2.5 seconds (instead of every 5 seconds). It also ignores network errors on the page with the form, since we don't want the network error modal to interrupt users who may be trying to type.