Make HP Actions for NYCHA users work properly (#1139)
I'm not super happy with this solution but it fixes #1113 without requiring a bunch of refactoring.

There are a few limitations with this approach.

On the front-end, we only show people the NYCHA main office address in HPA/EHPA's "your landlord" step if they explicitly said their lease type was "NYCHA" during onboarding.

On the back-end, Actually filling out the user's landlord info in the HPA filling process is slightly different and requires a bit more data-intensive work.  The general algorithm goes like this:

1. If the user explicitly declared that they have a NYCHA lease, _always_ sue the main NYCHA office, even if they don't appear to be in a NYCHA BBL.
2. Otherwise, if the user has manually entered landlord details (i.e., that weren't looked-up automatically at some point), sue that address.
3. Otherwise, see if the user's address has an HPD registration. If so, use that, and also sue any management company associated with the address.
4. Otherwise, see if the user's BBL is a NYCHA BBL, and if so, sue the main NYCHA office.  (Doing this after looking up HPD registration should hopefully ensure that we sue the right entity in the case of NYCHA RAD conversions).
5. Otherwise, don't fill in the landlord information at all.

The discrepancies between the front-end and back-end algorithms sometimes result in a few mismatches.  Most notably, if the user didn't say they were NYCHA but they are in a NYCHA BBL, we will show them their local NYCHA management office in the "your landlord" step but actually fill in the NYCHA main office in their papers.  While this is unfortunate, only a small minority of NYCHA people self-report as being non-NYCHA, and in the end it's still ultimately NYCHA that's being sued (just a different address) so I think it's OK for now, but I've filed #1140 to address it later.