Email Emergency HP actions to borough housing court (#1141)
Fixes #1103 by adding a singleton `Config` model to the `hpaction` app, which stores email addresses to all borough housing courts.  I didn't attempt to over-abstract this because we might eventually need to send certain HPs to e.g. Centers for Court Innovation in the future or somesuch.

An error is logged if no housing court is found for a user, since currently email is the only way for a case to actually be filed.