Add a confirmation modal to EHPA signing (#1150)
This adds a pre-signing modal prior to the signing process, and changes the copy of the pre-form generation and pre-signing steps to ensure the user understands what's about to happen, and how long it might take.  It also fixes a bug whereby following "back" on this step would take the user back to the required landlord details, rather than the optional ones.