Convert optional step 2 BooleanFields to NullBooleanFields (#1154)
Ugh, there appears to be a bug in graphene-Django whereby `BooleanField(null=True, blank=True)` are not converted to nullable booleans in the GraphQL query schema, but rather _non-nullable_ boolean fields.  This then causes an GraphQL validation error when such fields are queried and their value is null.

This fixes things by changing such fields to be `NullBooleanField` instances, which doesn't change the underlying DB schema or the behavior of the app, aside from the way graphene-Django introspects them and represents them as a GraphQL schema.

Finally, it also exposes one such field, `has_called_311`, and adds tests to ensure that it works as expected.