Animate moratorium banner entry/exit to reduce jank (#1164)
This adds a slide effect to the moratorium banner to reduce jank.

The rationale for this wasn't just to provide a nice effect, but also because when the page location changed to one that didn't have the banner, if a loading transition was occurring due to a JS bundle load, this caused the page contents to suddenly "jump up" the height of the banner, because it disappeared (as it wasn't within in the `<LoadingOverlayManager>` so it was getting the latest location props, rather than the location props of the transitioning-out page).  This was disorienting.

Aside from that, this also makes the banner use the `Routes` object instead of hard-coding the routes as strings, which will make the banner work if we add internationalization before COVID is over.  It also ensures the banner shows up on the Emergency HP splash page.