Validate email uniqueness during onboarding. (#1165)
This prevents new users from signing up with the addresses of existing users.

This does not actually get rid of #1053, as it doesn't actually enforce the uniqueness of email addresses at the database level (that will involve multiple migrations).  But at least it errs on the side of cautiousness and keeps our data clean, in case e.g. we ever want to allow users to log in via email.

Note that this doesn't necessarily make it harder to demo the product, since folks with Gmail and some other kinds of email accounts can use plus signs in their addresses to make their emails unique, while still having the emails delivered to their inbox.